+++ lua/CompMod/Marine/AdvancedWeapons/Post/Balance.lua
@@ -1,10 +1,10 @@
 -- Munitions includes shotgun and heavy machine gun
-kMunitionsTechResearchCost = 25
-kMunitionsTechResearchTime = 45
+kMunitionsTechResearchCost = 35
+kMunitionsTechResearchTime = 90
 
 -- Demolitions includes flamethrower and grenade launcher
 kDemolitionsTechResearchCost = 15
-kDemolitionsTechResearchTime = 45
+kDemolitionsTechResearchTime = 30
 
 kAdvancedArmoryUpgradeCost = 15
-kAdvancedArmoryResearchTime = 60
+kAdvancedArmoryResearchTime = 45