+++ lua/CompMod/Marine/Structures/RoboticsFactory/Cost/Post/Balance.lua
@@ -1,0 +1,2 @@
+kRoboticsFactoryCost = 5 --10
+kUpgradeRoboticsFactoryCost = 15 --5